Remove express deps for compression tests
